requirements.txt
flet==0.24.1
mysql-connector-python==9.1.0
werkzeug==3.1.0